Update README.md
Burp Suite Professional v2024.12.1

Burp Suite is a reliable and practical platform that provides you with a simple means of performing security testing of web applications. It gives you full control, letting you combine advanced manual techniques with various tools that seamlessly work together to support the entire testing process. The utility is easy-to-use and intuitive and does not require you to perform advanced actions in order to analyze, scan and exploit web apps. It is highly configurable and comes with useful features to assist experienced testers with their work.

[Download Burp Suite Professional v2024.12.1](https://developer.team/web-development/35222-burp-suite-professional-v2024121.html)